pyproject.toml
[tool.poetry]
name = "stegosaurus"
version = "0.1.0"
description = ""
authors = ["Naka <alex.naka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.107"
openai = "^0.27.2"
umap-learn = "^0.5.3"
scikit-learn = "^1.2.2"
seaborn = "^0.12.2"
pandas = "^1.5.3"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
